refactor: common toolbox
